% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annotations.R
\name{getAllAnnotationData}
\alias{getAllAnnotationData}
\title{Dowload a list of all annotations}
\usage{
getAllAnnotationData(path = "/R/annotations")
}
\arguments{
\item{path}{the view containing the annotations}
}
\value{
a \code{data.table}. The column \code{id} refers to the 
unique identifier of an annotation. \code{start} and \code{end} identify the
section of this annotation and \code{file} maps the URL of the raw audio file.
Other annotation data are available as extra columns.
}
\description{
Download the information corresponding 
to all available annotaions on bioacoustica
}

