% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annotations.R
\name{getAnnotationFile}
\alias{getAnnotationFile}
\title{Fetch a portion of a remote wav file}
\usage{
getAnnotationFile(file, start, end, dst, force = F, first_chunk = 1024,
  verbose = F)
}
\arguments{
\item{file}{the URL of a file}

\item{start}{the start of the desired section, in second}

\item{end}{the end of the desired section, in second}

\item{dst}{a character vector indicating where to save the resulting file}

\item{force}{whether to downoad and overwrite pre-existing file}

\item{first_chunk}{the number of bytes downloaded at first in order to identify the file}

\item{verbose}{weher to print extra information}
}
\value{
the path to resulting file
}
\description{
This function can download only a portion of a wave file
stored online and store it locally as a new wave file.
It proceeds by dowloading metadata, uses them to locate the file chunk
and reconstitute a file
}
\seealso{
\code{\link{dowloadFilesForAnnotations}}, a wrapper around this function
}

