% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annotations.R
\name{dowloadFilesForAnnotations}
\alias{dowloadFilesForAnnotations}
\title{Retreive list of annotation files from a query}
\usage{
dowloadFilesForAnnotations(query, dst_dir, prefix = "annotation")
}
\arguments{
\item{query}{a data.table or data.frame (see details)}

\item{dst_dir}{a directory to store the resulting data}

\item{prefix}{a string to prepend to filenames}
}
\description{
This function uses a formated query
(i.e. table) to retreive annotation raw data
}
\details{
\code{query} must contain the columns \code{id},
\code{start}, \code{end}, \code{file}. They indicate, the annotation uid
, the start and end of the wav file, and the url of the wav file, respactively.
Typically, it will be obtained through \code{\link{getAllAnnotationData}}.
}
\examples{
\dontrun{
all_annotations = getAllAnnotationData()
query = all_annotations[author=="qgeissmann"]
my_annotations <- dowloadFilesForAnnotations(query, 
                               dst_dir = "/tmp/my_annotation")
}
}
\seealso{
\code{\link{getAllAnnotationData}}
}

