% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annotations.R
\name{dowloadFilesForAnnotations}
\alias{dowloadFilesForAnnotations}
\title{Retreive list of annotation files from a query}
\usage{
dowloadFilesForAnnotations(query, dst_dir, prefix = "annotation", ...)
}
\arguments{
\item{query}{a data.table or data.frame (see details)}

\item{dst_dir}{a directory to store the resulting data}

\item{prefix}{a string to prepend to filenames}

\item{...}{extra argument passed to \code{link{getAnnotationFile}}}
}
\description{
This function uses a formated query
(i.e. table) to retreive annotation raw data
}
\details{
\code{query} must contain the columns \code{id},
\code{start}, \code{end}, \code{file}. They indicate, the annotation uid
, the start and end of the wav file, and the url of the wav file, respactively.
Typically, it will be obtained through \code{\link{getAllAnnotationData}}.
}
\examples{
\dontrun{
# we retreive all annotations from the database
all_annotations = getAllAnnotationData()
# we filter for only those made by one author before a certain date
query = all_annotations[author=="qgeissmann" & date < "2016-10-26"]
# we make one file for each annotation and save it in dst_dir
my_annotations <- dowloadFilesForAnnotations(query,
                               dst_dir = "/tmp/my_annotation",
                               verbose=T)
}
}
\seealso{
\code{\link{getAllAnnotationData}}
}

